Fix indicators listing on helper and add js tests
Due to a wrong logic implementation, the <ol> tag
was being provided with <ul> tags as children, which
is NOT the correct way to do so. This was fixed
and a default initial indicator as highlighted was also
added. Jasmine specs were added to assure this feature
will work as expected.